+++ src/wallet/core/mod.rs
@@ -4,3 +4,4 @@ pub mod imports;
 pub mod message;
 pub mod tx;
 pub mod utils;
+pub mod utxo;